refactor(canopee): pass Icon as prop to Toggle via Apollo/LF wrappers
