+++ packages/canopee-react/src/prospect-client/Toggle/ToggleCommon.tsx
@@ -1,8 +1,12 @@
 import check from "@material-symbols/svg-400/rounded/check.svg";
 import close from "@material-symbols/svg-400/rounded/close.svg";
-import { forwardRef, type InputHTMLAttributes, useId } from "react";
-
-import { Icon } from "../Icon/IconCommon";
+import {
+  type ComponentType,
+  forwardRef,
+  type InputHTMLAttributes,
+  useId,
+} from "react";
+import type { IconProps } from "../Icon/IconCommon";
 import { getComponentClassName } from "../utilities/getComponentClassName";
 
 export type ToggleProps = Omit<
@@ -10,8 +14,12 @@ export type ToggleProps = Omit<
   "style" | "type"
 >;
 
-export const Toggle = forwardRef<HTMLInputElement, ToggleProps>(
-  ({ className, ...inputProps }, ref) => {
+export type ToggleCommonProps = ToggleProps & {
+  IconComponent: ComponentType<IconProps>;
+};
+
+export const ToggleCommon = forwardRef<HTMLInputElement, ToggleCommonProps>(
+  ({ className, IconComponent, ...inputProps }, ref) => {
     let inputId = useId();
     inputId = inputProps.id ?? inputId;
 
@@ -22,13 +30,13 @@ export const Toggle = forwardRef<HTMLInputElement, ToggleProps>(
       >
         <div className="af-toggle__root">
           <span className="af-toggle__handle">
-            <Icon
+            <IconComponent
               src={check}
               aria-hidden="true"
               className="af-toggle__icon af-toggle__icon--check"
               size="XS"
             />
-            <Icon
+            <IconComponent
               src={close}
               aria-hidden="true"
               className="af-toggle__icon af-toggle__icon--close"
@@ -41,4 +49,4 @@ export const Toggle = forwardRef<HTMLInputElement, ToggleProps>(
     );
   },
 );
-Toggle.displayName = "Toggle";
+ToggleCommon.displayName = "ToggleCommon";
